progress towards windows port
Primary changes were removing usage of makeWin32Path as in most cases we do not need a full path to what we are working with and I guess I broke it to the point where it causes heap corruption and I do not care to fix it considering I don't want to use it. Output from processes is now correctly read. I've also decided to get rid of support for pseudo terminals entirely for now since getting them work between win32 and linux has proven to be difficult and is an unecessary roadblock for getting this all working. However I have not implemented its removal yet and will shortly. For now on win32 at least when we try to create a PTY we just redirect to creating a normal process, but this only captures stdout for now, which is a problem because things like to output to stderr instead.

Lpp still needs to be properly ported (not allowed to apply __declspec(dllexport) to multiple declaration on windows apparently) and so once that is fixed and whatever else needs done with it, we should be able to move on to porting ECS.

I'm also checking in .clangd so that we can have consistent clang-tidy warnings across systems for anyone using clangd.